changed to complete setup
