+++ bin/functions/video.js
@@ -1,8 +1,65 @@
 // VIDEO LOADER
 
+/*
+document.querySelector( '#button-downloader' ).onclick = function(){
+    fetch("https://videoboothfiles.blob.core.windows.net/videos/1zu18647161806378.mp4")
+    .then(resp => resp.blob())
+    .then(blob => {
+        const url = window.URL.createObjectURL(blob);
+        const a = document.createElement("a");
+        a.style.display = "none";
+        a.href = url;
+        // the filename you want
+        a.download = "video.mp4";
+        document.body.appendChild(a);
+        a.click();
+        window.URL.revokeObjectURL(url);
+        alert("your file has downloaded!"); // or you know, something with better UX...
+    })
+    .catch(() => alert("oh no!"));
+};
+*/
+
 // import modules
 import { json as config } from '../config.js';
 import { checkVideo } from './checks.js';
+
+function addButtons(){
+    document.querySelector( '#wrap-button-video-1zu1' ).onclick = function(){
+        fetch(config.video.blob + '1zu1' + id + '.' + config.video.filetype)
+        .then(resp => resp.blob())
+        .then(blob => {
+            const url = window.URL.createObjectURL(blob);
+            const a = document.createElement("a");
+            a.style.display = "none";
+            a.href = url;
+            // the filename you want
+            a.download = "1zu1video.mp4";
+            document.body.appendChild(a);
+            a.click();
+            window.URL.revokeObjectURL(url);
+            alert("File is ready to download!"); // or you know, something with better UX...
+        })
+        .catch(() => alert("oh no!"));
+    };
+    document.querySelector( '#wrap-button-video' ).onclick = function(){
+        fetch(config.video.blob + '1zu1' + id + '.' + config.video.filetype)
+        .then(resp => resp.blob())
+        .then(blob => {
+            const url = window.URL.createObjectURL(blob);
+            const a = document.createElement("a");
+            a.style.display = "none";
+            a.href = url;
+            // the filename you want
+            a.download = "video.mp4";
+            document.body.appendChild(a);
+            a.click();
+            window.URL.revokeObjectURL(url);
+            alert("File is ready to download!"); // or you know, something with better UX...
+        })
+        .catch(() => alert("oh no!"));
+    };
+}
 
 // make video player / download buttons visible
 function loadVideo( id ){
@@ -11,19 +68,20 @@ function loadVideo( id ){
 
     document.querySelector( '.elementor-video' ).src = config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token;
     var video1zu1 = document.getElementById( 'player-video-1zu1' );
-    var button1zu1 = document.getElementById( 'wrap-button-video-1zu1' );
+    //var button1zu1 = document.getElementById( 'wrap-button-video-1zu1' );
     video1zu1.classList.add( 'video-accessable' );
-    button1zu1.classList.add( 'video-accessable' );
-    //video1zu1.innerHTML = video1zu1.innerHTML.replace( 'https://videoboothfiles/', config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token );
-    button1zu1.innerHTML = button1zu1.innerHTML.replace( 'http://videolink1zu1', config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token );
-    var statusVideo = checkVideo();
+    //button1zu1.classList.add( 'video-accessable' );
+    video1zu1.innerHTML = video1zu1.innerHTML.replace( 'https://videoboothfiles/', config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token );
+    //button1zu1.innerHTML = button1zu1.innerHTML.replace( 'http://videolink1zu1', config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token );
+    /* var statusVideo = checkVideo();
     if ( statusVideo == true ){
         console.log( 'WEB APP => making video button 2 visible ⬇️' );
 
-        var video = document.getElementById( 'wrap-button-video' );
+        //var video = document.getElementById( 'wrap-button-video' );
         //video.innerHTML = video.innerHTML.replace( 'http://videolink', config.video.blob + id + '.' + config.video.filetype + '?' + config.video.token );
-        video.classList.add( 'video-accessable' );
-    }
+        //video.classList.add( 'video-accessable' );
+    } */
+    addButtons();
 }
 
 // change video urls
@@ -31,12 +89,14 @@ function switchLink( id ){
     console.log( 'WEB APP => changing video urls 🔗' );
 
     document.querySelector( '.elementor-video' ).src = config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token;
+    
+    
     //var video1zu1 = document.getElementById( 'player-video-1zu1' );
-    var button1zu1 = document.getElementById( 'wrap-button-video-1zu1' );
-    var video = document.getElementById( 'wrap-button-video' );
+    //var button1zu1 = document.getElementById( 'wrap-button-video-1zu1' );
+    //var video = document.getElementById( 'wrap-button-video' );
     //video1zu1.innerHTML = video1zu1.innerHTML.replace( 'https://videoboothfiles/', config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token );
-    button1zu1.innerHTML = button1zu1.innerHTML.replace( 'http://videolink1zu1', config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token );
-    video.innerHTML = video.innerHTML.replace( 'http://videolink', config.video.blob + id + '.' + config.video.filetype + '?' + config.video.token );
+    //button1zu1.innerHTML = button1zu1.innerHTML.replace( 'http://videolink1zu1', config.video.blob + '1zu1' + id + '.' + config.video.filetype + '?' + config.video.token );
+    //video.innerHTML = video.innerHTML.replace( 'http://videolink', config.video.blob + id + '.' + config.video.filetype + '?' + config.video.token );
 }
 
 // export modules
